refactor: ディレクトリ階層構造の変更(#6)
