+++ src/components/common/QuizCard/QuizCard.tsx
