+++ src/components/common/QuizCard/index.ts
@@ -1,0 +1,3 @@
+import QuizCard from './QuizCard';
+
+export default QuizCard;
