+++ src/components/model/Question/QuestionAnswerButton/index.ts
@@ -1,0 +1,3 @@
+import QuestionAnswerButton from './QuestionAnswerButton';
+
+export default QuestionAnswerButton;
