+++ src/components/page/Home/index.ts
@@ -1,0 +1,3 @@
+import HomePage from './Home.page';
+
+export default HomePage;
