+++ src/components/page/regions/Quiz/Quiz.page.tsx
@@ -1,0 +1,13 @@
+import { VFC } from 'react';
+import Layout from '@/components/common/Layout/Layout';
+import Quiz from './Quiz';
+
+const QuizPage: VFC = () => {
+  return (
+    <Layout>
+      <Quiz />
+    </Layout>
+  );
+};
+
+export default QuizPage;
