+++ src/components/page/regions/Quiz/index.ts
@@ -1,0 +1,3 @@
+import QuizPage from './Quiz.page';
+
+export default QuizPage;
