+++ src/pages/regions/[region].tsx
@@ -1,0 +1,7 @@
+import RegionQuiz from '@/components/page/regions/Quiz';
+
+const Quiz = () => {
+  return <RegionQuiz />;
+};
+
+export default Quiz;
